+++ src/main/java/com/goufn/permission/service/RoleService.java
@@ -1,12 +1,12 @@
 package com.goufn.permission.service;
 
-import com.goufn.permission.entity.SysRole;
+import com.goufn.permission.model.SysRole;
 
 import java.util.Set;
 
 public interface RoleService {
 
-    Set<String> findRoleByUserId(int userId);
+    Set<String> findRoleByUserId(long userId);
 
     SysRole findByName(String roleName);
