+++ src/main/java/com/goufn/permission/service/impl/RoleServiceImpl.java
@@ -2,10 +2,9 @@ package com.goufn.permission.service.impl;
 
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
-import com.goufn.permission.entity.SysRole;
+import com.goufn.permission.model.SysRole;
 import com.goufn.permission.mapper.RoleMapper;
 import com.goufn.permission.service.RoleService;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.Set;
@@ -14,7 +13,7 @@ import java.util.Set;
 public class RoleServiceImpl extends ServiceImpl<RoleMapper, SysRole> implements RoleService {
 
     @Override
-    public Set<String> findRoleByUserId(int userId) {
+    public Set<String> findRoleByUserId(long userId) {
         return baseMapper.findRoleByUserId(userId);
     }
